adding profile image

+++ todo/views.py
@@ -1,7 +1,7 @@
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.views.decorators.csrf import csrf_exempt
-
+from django.contrib.auth.decorators import login_required
 from todo.models import Todo
 
 # Create your views here.
@@ -15,6 +15,7 @@ def index(request):
     else:
         return HttpResponse("Invalid request method", setatus=405)
 
+@login_required(login_url='/user/login/')
 @csrf_exempt
 def create(request):
     if request.method == "POST":
@@ -37,48 +38,25 @@ def read(request, todo_id):
 def delete(request, todo_id):
     if request.method == "POST":
         todo = Todo.objects.get(id=todo_id)
-        todo.delete()
-        return redirect("/todo/")
+        if request.user == todo.user:
+            todo.delete()
+            return redirect("/todo/")
+        else:
+            return HttpResponse("Invalid request method", setatus=403)
+
     else:
         return HttpResponse("Invalid request method", setatus=405)
 
 @csrf_exempt     
 def update(request, todo_id):
     if request.method == "POST":
-        todo = Todo.objects.get(id=todo_id)
-        todo.content = request.POST["content"]
-        todo.save()
-        return redirect(f"/todo/{todo_id}")
-
-        # return redirect("/todo/")
-    elif request.method =="GET":
-        todo = Todo.objects.get(id=todo_id)
-        context = {
-            "todo": todo,
-        }
-        return render(request, 'todo/update.html', context)
-    else:
-        return HttpResponse("Invalid request method", setatus=405)
-    
-
-@csrf_exempt 
-def delete(request, todo_id):
-    if request.method == "POST":
-        todo = Todo.objects.get(id=todo_id)
-        todo.delete()
-        return redirect("/todo/")
-    else:
-        return HttpResponse("Invalid request method", setatus=405)
-
-@csrf_exempt     
-def update(request, todo_id):
-    if request.method == "POST":
-        todo = Todo.objects.get(id=todo_id)
-        todo.content = request.POST["content"]
-        todo.save()
-        return redirect(f"/todo/{todo_id}")
-
-        return redirect("/todo/")
+        if request.user == todo.user:
+            todo = Todo.objects.get(id=todo_id)
+            todo.content = request.POST["content"]
+            todo.save()
+            return redirect(f"/todo/{todo_id}")
+        else:
+            return HttpResponse("Invalid request method", setatus=403)
     elif request.method =="GET":
         todo = Todo.objects.get(id=todo_id)
         context = {
